Encapsulate the insert tag related methods in a separate class
